docker_cassandra/requirements.txt
cassandra-driver==2.6.0
Flask==0.10.1
futures==3.0.3
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
six==1.9.0
Werkzeug==0.10.4
